requirements.txt
streamlit==0.65.2
numpy==1.18.5
opencv_python==4.4.0.42
Pillow==7.2.0
tensorflow-cpu==2.3.0
